Todo transactional

+++ src/main/java/com/java/TrainningJV/services/impl/ProductServiceImpl.java
@@ -1,9 +1,19 @@
 package com.java.TrainningJV.services.impl;
 
+import java.util.List;
+
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import com.java.TrainningJV.dtos.request.ProductRequest;
-import com.java.TrainningJV.mappers.ProductMapper;
+import com.java.TrainningJV.dtos.response.ProductPageResponse;
+import com.java.TrainningJV.dtos.response.ProductResponse;
+import com.java.TrainningJV.exceptions.BadRequestException;
+import com.java.TrainningJV.exceptions.ResourceNotFoundException;
+import com.java.TrainningJV.mappers.mapper.CategoryMapper;
+import com.java.TrainningJV.mappers.mapper.ProductMapper;
+import com.java.TrainningJV.mappers.mapperCustom.ProductMapperCustom;
+import com.java.TrainningJV.models.Category;
 import com.java.TrainningJV.models.Product;
 import com.java.TrainningJV.services.ProductService;
 
@@ -15,6 +25,28 @@ import lombok.extern.slf4j.Slf4j;
 @RequiredArgsConstructor
 public class ProductServiceImpl implements ProductService {
     private final ProductMapper productMapper;
+    private final ProductMapperCustom productMapperCustom;
+    private final CategoryMapper categoryMapper;
+
+    @Override
+    public ProductPageResponse getAllProducts(int page, int size) {
+        log.info("Getting all products with paging: page{} size{}", page, size);
+        int offset = (page - 1 ) * size;
+        List<ProductResponse> productPage = productMapperCustom.getAllProducts(offset, size);
+
+        int totalElements = productMapperCustom.countProduct();
+        int totalPages = (int) Math.ceil((double)totalElements/size);
+        ProductPageResponse productPageResponse = new ProductPageResponse();
+        productPageResponse.setPageNumber(page);
+        productPageResponse.setPageSize(size);
+        productPageResponse.setTotalElements(totalElements);
+        productPageResponse.setTotalPages(totalPages);
+        productPageResponse.setProduct(productPage);
+
+        log.info("get product all successfull");
+
+        return productPageResponse;
+    }
 
     @Override
     public Product getProductById(Integer id) {
@@ -22,29 +54,38 @@ public class ProductServiceImpl implements ProductService {
         Product exitingProduct = productMapper.selectByPrimaryKey(id);
         if (exitingProduct == null) {
             log.info("Product with id: " + id + " not found");
-            return null;
+            throw new ResourceNotFoundException("Product", "id", id);
         }
-        log.info("Product with id: {} ", id );
 
+        log.info("Product with id: {} ", id );
         return exitingProduct;
     }
 
     @Override
-    public Integer addProduct(ProductRequest productRequest) {
+    @Transactional
+    public Product addProduct(ProductRequest productRequest) {
         log.info("Add product: {} ", productRequest);
-       
-        Product product = Product.builder()
+        
+        Category exitingCategory = categoryMapper.selectByPrimaryKey(productRequest.getCategoryId());
+        if(exitingCategory == null){
+            log.info("Category with id: {} not found", productRequest.getCategoryId());
+            throw new ResourceNotFoundException("Category", "id", productRequest.getCategoryId());
+        }
+        
+        Product newProduct = Product.builder()
                 .name(productRequest.getNameProduct())
                 .price(productRequest.getPrice())
                 .description(productRequest.getDescription())
                 .stockQuantity(productRequest.getStockQuantity())
+                .categoryId(productRequest.getCategoryId())
                 .build();
-        int result = productMapper.insert(product);  
+        int result = productMapper.insert(newProduct);  
         if(result > 0) {
-            log.info("Product created successfully: {}", product);
-            return product.getId(); // Assuming the ID is set after creation
+            log.info("Product created successfully: {}", newProduct);
+            Product product = productMapper.selectByPrimaryKey(newProduct.getId());
+            return product;
         } else {
-            throw new RuntimeException("Failed to create product");
+            throw new BadRequestException("Failed to create product");
         }
     
     }
@@ -52,25 +93,23 @@ public class ProductServiceImpl implements ProductService {
     @Override
     public void deleteProduct(Integer id) {
         log.info("Delete product: {} ", id);
-        // Product exitingProduct = productMapper.selectByPrimaryKey(id);
-        productMapper.deleteByPrimaryKey(id);
-        // nt result = productMapper.deleteByPrimaryKey(id);
-        // if (result > 0) {
-        //     log.info("Product deleted successfully");
-        // }else {
-        //     throw new RuntimeException("Product delete failed");
-        // }
+        Product exitingProduct = productMapper.selectByPrimaryKey(id);
 
+        if(exitingProduct == null){
+            log.info("Product not found: {}", id);
+            throw new ResourceNotFoundException("Product", "id", id);
+        }
+        productMapper.deleteByPrimaryKey(id);
     }
 
     @Override
-    public int updateProduct(Integer id, ProductRequest productRequest) {
+    @Transactional
+    public Product updateProduct(Integer id, ProductRequest productRequest) {
         log.info("Update product with id: {} and request: {}", id, productRequest);
-        // Product existingProduct = productMapper.selectByPrimaryKey(id);
-        // log.info("Existing product: {}", existingProduct);
-        // if (existingProduct == null) {
-        //     throw new RuntimeException("Product not found with id: " + id);
-        // }
+        Product existingProduct = productMapper.selectByPrimaryKey(id);
+        if (existingProduct == null) {
+           throw new ResourceNotFoundException("Product", "id", id);
+        }
         
         Product updatedProduct = Product.builder()
                 .id(id)
@@ -81,6 +120,13 @@ public class ProductServiceImpl implements ProductService {
                 .build();
         
         log.info("Updated product: {}", updatedProduct);
-        return productMapper.updateByPrimaryKey(updatedProduct);
+        int rows = productMapper.updateByPrimaryKey(updatedProduct);
+        if(rows != 1) {
+            log.error("Failed to update product with id: {}", id);
+            throw new BadRequestException("Failed to update product");
+        }
+        Product productUpadate = productMapper.selectByPrimaryKey(id);
+        log.info("Product updated successfully: {}", productUpadate);
+        return productUpadate;
     }
 }
